emails funcionando

+++ billeteravirtual/src/main/java/ar/com/ada/api/billeteravirtual/models/request/TransferenciaRequest.java
@@ -1,11 +1,13 @@
 package ar.com.ada.api.billeteravirtual.models.request;
 
+import java.math.BigDecimal;
+
 /**
  * TransferenciaRequest
  */
 public class TransferenciaRequest {
 
     public String emailUsuarioDest;
-    public double importe;
+    public BigDecimal importe;
     public String concepto;
 }
