+++ billeteravirtual/src/main/java/ar/com/ada/api/billeteravirtual/models/response/SaldoResponse.java
@@ -1,5 +1,7 @@
 package ar.com.ada.api.billeteravirtual.models.response;
 
+import java.math.BigDecimal;
+
 /**
  * SaldoResponse
  */
@@ -7,5 +9,5 @@ public class SaldoResponse {
 
     public int billeteraId;
     public String moneda;
-    public double saldo;
+    public BigDecimal saldo;
 }
